requirements.txt
python-telegram-bot==21.6
python-docx==1.1.2
aiofiles==23.2.1
python-dotenv==1.0.1
fpdf2==2.7.9
